python-poetry-app/pyproject.toml
[tool.poetry]
name = "python-poetry-crypto-app"
version = "1.0.0"
description = "Sample Poetry project with crypto and non-crypto dependencies"
authors = ["Example Author <author@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"

# Crypto dependencies
cryptography = ">=42.0.0,<44.0.0"
pyopenssl = "^24.2.1"
paramiko = "^3.4.0"
bcrypt = "^4.1.0"
pyjwt = {version = "^2.8.0", extras = ["crypto"]}
pynacl = "^1.5.0"

# Non-crypto dependencies
requests = {version = "^2.31.0", extras = ["socks"]}
pydantic = "^2.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.1.0"
isort = "^5.13.0"
mypy = "^1.8.0"
types-requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100
